refactor(core): split module_resolver/mod.rs into loader/expander/typecheck/resolver
Split the 3,447-line `carina-core/src/module_resolver/mod.rs` into
focused submodules per the issue's target layout:

- `error.rs`     — `ModuleError`
- `loader.rs`    — `load_module`, `load_directory_module`,
                   `load_module_from_directory`, `get_parsed_file`,
                   `derive_module_name`, `sorted_crn_paths_in`
- `typecheck.rs` — `check_module_arg_type`, `TypeCheckResult`,
                   `check_type_match`
- `expander.rs`  — `ModuleResolver::expand_module_call`,
                   `substitute_arguments`, `rewrite_intra_module_refs`,
                   `instance_prefix_for_call`,
                   `parse_synthetic_instance_prefix`,
                   `split_instance_prefix`,
                   `reconcile_anonymous_module_instances`,
                   `rewrite_ref_prefixes`
- `resolver.rs`  — `ModuleResolver` struct + impl, `resolve_modules`,
                   `resolve_modules_with_config`
- `tests.rs`     — the existing test module (mechanically extracted)
- `mod.rs`       — facade re-exports only (~46 lines)

Mechanical move: no logic change, no public API change. Cross-module
fields on `ModuleResolver` are `pub(super)` so `expander.rs` can reach
them without exposing them outside `module_resolver`.

The loader gains directory-scoped acceptance tests using
`tempfile::tempdir()` with multiple `.crn` siblings, per the
"directory-scoped, never single-file" project rule.

Verification:
- `cargo test -p carina-core module_resolver::` — 55 passed
- `cargo test -p carina-core` — 1385 unit + 8 doctests pass
- `cargo test -p carina-cli` — 264 passed
- `cargo test -p carina-lsp` — all green
- `cargo clippy --workspace --all-targets -- -D warnings` — clean
- `cargo fmt --check` — clean
- `carina validate carina-rs/infra/aws/management/github-oidc`
  (real-infra smoke with module import) — passes

Closes #2264
